# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ai-social-agent Operator API",
    "version": "1.0.0",
    "description": "Localhost operator console API (Wave-3 envelope: error, code, requestId)."
  },
  "servers": [{ "url": "http://127.0.0.1:8787" }],
  "paths": {
...
    "/api/publish/history": {
      "get": {
        "summary": "Publish history",
        "responses": { "200": { "description": "ok" }, "400": { "description": "bad limit" } }
      }
    },
...
  }
}